Rename :data:users:repository module to :data:users:core
As this module includes :data:users:model via api() I think it's better to call it :core than :repository.
