+++ data/users/core/src/main/kotlin/dev/burnoo/demo/listapp/data/users/core/di/Module.kt
@@ -1,8 +1,8 @@
-package dev.burnoo.demo.listapp.data.users.repository.di
+package dev.burnoo.demo.listapp.data.users.core.di
 
+import dev.burnoo.demo.listapp.data.users.core.NetworkUsersRepository
+import dev.burnoo.demo.listapp.data.users.core.UsersRepository
 import dev.burnoo.demo.listapp.data.users.network.di.usersNetworkModule
-import dev.burnoo.demo.listapp.data.users.repository.NetworkUsersRepository
-import dev.burnoo.demo.listapp.data.users.repository.UsersRepository
 import org.koin.dsl.module
 
 val usersRepositoryModule = module {
