+++ data/users/core/src/main/kotlin/dev/burnoo/demo/listapp/data/users/core/mappers/User.kt
@@ -1,4 +1,4 @@
-package dev.burnoo.demo.listapp.data.users.repository.mappers
+package dev.burnoo.demo.listapp.data.users.core.mappers
 
 import dev.burnoo.demo.listapp.data.users.model.User
 import dev.burnoo.demo.listapp.data.users.network.model.NetworkUser
