+++ ui/userlist/src/main/java/dev/burnoo/demo/listapp/userlist/di/Module.kt
@@ -1,6 +1,6 @@
 package dev.burnoo.demo.listapp.userlist.di
 
-import dev.burnoo.demo.listapp.data.users.repository.di.usersRepositoryModule
+import dev.burnoo.demo.listapp.data.users.core.di.usersRepositoryModule
 import dev.burnoo.demo.listapp.userlist.UserListViewModel
 import org.koin.androidx.viewmodel.dsl.viewModelOf
 import org.koin.dsl.module
